updated the db setup script for seeding

+++ application/lib/db.js
@@ -3,11 +3,25 @@
  */
 'use strict';
 let mongoose = require('mongoose'),
+  _ = require('lodash'),
   async = require('async'),
   config = require('config'),
-  dbSetup = require('../../db/db-setup-script');
+  dbSeedRecords = require('../../db/db-setup-script'),
+  utils = require('../lib/utils');
 let connection;
 let connected = false;
+let dbCreateIndex = true;
+let dbSeedCollection = true;
+let dbDropCollection = true;
+let dbIgnoreForDropCollection = true;
+let dbIgnoreDropCollections = [];
+let dbIgnoreForSeedCollection = true;
+let dbIgnoreSeedCollections = ['payments', 'reservations', 'travelers'];
+let dbSeedCollections = ['aircrafts', 'airports', 'flights', 'inventory', 'payments', 'reservations', 'travelers', 'users'];
+let dbIgnoreForIndexCollection = true;
+let dbIgnoreIndexCollections = ['payments', 'reservations', 'travelers'];
+let dbIndexCollections = ['aircrafts', 'airports', 'flights', 'inventory', 'payments', 'reservations', 'travelers', 'users'];
+var hashPassword;
 
 async function connectToMongo() {
   logger.info('METHOD ENTRY - application.lib.db.connectToMongo');
@@ -17,7 +31,7 @@ async function connectToMongo() {
   await mongoose.set('useFindAndModify', false);
   await mongoose.set('useCreateIndex', true);
   await mongoose.set('autoIndex', false);
-  await mongoose.connect(config.get('database'), {useNewUrlParser: true}).
+  await mongoose.connect(config.get('database'), { useNewUrlParser: true }).
     catch(error => {
       logger.error('MongoDb connection error: ' + err);
       return error;
@@ -26,7 +40,6 @@ async function connectToMongo() {
   logger.info('METHOD EXIT - application.lib.db.connectToMongo');
   return connected;
 }
-
 async function closeConnection() {
   logger.info('METHOD ENTRY - application.lib.db.closeConnection');
   await mongoose.disconnect();
@@ -34,41 +47,42 @@ async function closeConnection() {
   logger.info('METHOD EXIT - application.lib.db.closeConnection');
   return connected;
 }
-
 async function setupDb() {
   await logger.info('METHOD ENTRY - application.lib.db.setupDb');
   connection = await mongoose.connection;
   if (!connection) {
     await logger.error('DB Connection promise not passed to setup DB');
+    return;
   }
   await logger.info('=========== Starting:  master records Patch ==============');
 
   async.waterfall([
-    async function(cb) {
-      await removeCollections(function(err) {
+    async function (cb) {
+      await removeCollections(function (err) {
         if (err) {
           return cb('Remove collections failed: ' + err);
         }
         return cb();
       });
     },
-    async function(cb) {
-      await createIndex('aircrafts', 'aircraft_no', function(err) {
+    async function (cb) {
+      await createIndexes(function (err) {
         if (err) {
           return cb('Create index failed: ' + err);
         }
         return cb();
       });
     },
-    async function(cb) {
-      await updateCollection('aircrafts', dbSetup.aircraftRecords, function(err) {
+    async function (cb) {
+      await updateCollections(function (err) {
         if (err) {
-          return cb('Update collection failed ' + err);
+          return cb('Update collections failed ' + err);
         }
         return cb();
       });
-    }],
-    async function(err) {
+    }
+  ],
+    async function (err) {
       if (err) {
         await logger.error('Master records Patch failed with error:- ' + err);
         return err;
@@ -78,13 +92,23 @@ async function setupDb() {
       return;
     });
 }
-
 async function removeCollections() {
   await logger.info('METHOD ENTRY - application.lib.db.removeCollections');
 
   const collections = await mongoose.connection.db.collections();
   for (let collection of collections) {
-    await collection.drop();
+    if (dbIgnoreForDropCollection) {
+      let result = _.indexOf(dbIgnoreDropCollections, collection.s.name);
+      if (result === -1) {
+        await logger.info('Collection ' + collection.s.name + ' is now dropped');
+        await collection.drop();
+      } else {
+        await logger.info('Collection ' + collection.s.name + ' drop is SKIPPED');
+      }
+    } else {
+      await logger.info('Collection ' + collection.s.name + ' is now dropped');
+      await collection.drop();
+    }
   }
   await logger.info('METHOD EXIT - application.lib.db.removeCollections');
 }
@@ -123,9 +147,9 @@ async function createIndex(collectionName, index_field, cb) {
     }
     // Create an index on the field
     collection.createIndex(
-      {[index_field]: 1},
-      {unique: true, background: true}
-      , function(err, indexName) {
+      { [index_field]: 1 },
+      { unique: true, background: true }
+      , function (err, indexName) {
         if (err) {
           logger.error('FAIL: Error occured while adding ' + indexName + '  to ' + collectionName + ' collection.')
           return err;
@@ -133,6 +157,54 @@ async function createIndex(collectionName, index_field, cb) {
       });
   });
   await logger.info('METHOD EXIT - application.lib.db.createIndex');
+}
+async function createIndexes(cb) {
+  await logger.info('METHOD ENTRY - application.lib.db.createIndexes');
+  _.forEach(dbIndexCollections, async (value) => {
+    if (dbIgnoreForIndexCollection) {
+      let result = _.indexOf(dbIgnoreIndexCollections, value);
+      if (result === -1) {
+        await logger.info('Collection ' + value + ' is now indexed');
+        mongoose.connection.db.collection(value, async (err, collection) => {
+          if (err) {
+            await logger.error('FAIL: Error in dropping ' + value + ' collection.');
+            return err;
+          }
+          // Create an index on the field
+          collection.createIndex(
+            { [dbSeedRecords.createIndex[value]]: 1 },
+            { unique: true, background: true }
+            , function (err, indexName) {
+              if (err) {
+                logger.error('FAIL: Error occured while adding ' + indexName + '  to ' + value + ' collection.')
+                return err;
+              }
+            });
+        });
+      } else {
+        await logger.info('Collection ' + value + ' indexing is SKIPPED');
+      }
+    } else {
+      await logger.info('Collection ' + value + ' is now indexed');
+      mongoose.connection.db.collection(value, async (err, collection) => {
+        if (err) {
+          await logger.error('FAIL: Error in indexing ' + value + ' collection.');
+          return err;
+        }
+        // Create an index on the field
+        collection.createIndex(
+          { [dbSeedRecords.createIndex[value]]: 1 },
+          { unique: true, background: true }
+          , function (err, indexName) {
+            if (err) {
+              logger.error('FAIL: Error occured while adding ' + indexName + '  to ' + value + ' collection.')
+              return err;
+            }
+          });
+      });
+    }
+  });
+  await logger.info('METHOD EXIT - application.lib.db.createIndexes');
 }
 async function updateCollection(collectionName, data, cb) {
   await logger.info('METHOD ENTRY - application.lib.db.updateCollection');
@@ -151,22 +223,81 @@ async function updateCollection(collectionName, data, cb) {
       await logger.error('FAIL: Error in dropping ' + collectionName + ' collection.');
       return err;
     }
-    async.each(data, function(obj, asynCb) {
-      collection.insertOne(obj, function(err) {
+    async.each(data, function (obj, asynCb) {
+      collection.insertOne(obj, function (err) {
         if (err) {
           logger.error('FAILED while resetting "' + collectionName + '" collection. Error: ' + err);
           return asynCb(err);
         }
         return asynCb();
       });
-    }, function(err) {
+    }, function (err) {
       if (err) {
         logger.error('FAIL: Error occured while adding records to ' + collectionName + ' collection.')
         return err;
       }
+      logger.info('Update collection name ' + collectionName + ' successful.');
     });
   });
   await logger.info('METHOD EXIT - application.lib.db.updateCollection');
+}
+async function updateCollections() {
+  await logger.info('METHOD ENTRY - application.lib.db.updateCollections');
+  _.forEach(dbSeedCollections, async (value) => {
+    if (dbIgnoreForSeedCollection) {
+      let result = _.indexOf(dbIgnoreSeedCollections, value);
+      if (result === -1) {
+        await logger.info('Collection ' + value + ' is now updated');
+        mongoose.connection.db.collection(value, async (err, collection) => {
+          if (err) {
+            await logger.error('FAIL: Error in updating ' + value + ' collection.');
+            return err;
+          }
+          async.each(dbSeedRecords[value], function (obj, asynCb) {
+            collection.insertOne(obj, function (err) {
+              if (err) {
+                logger.error('FAILED while updating "' + value + '" collection. Error: ' + err);
+                return asynCb(err);
+              }
+              return asynCb();
+            });
+          }, function (err) {
+            if (err) {
+              logger.error('FAIL: Error occured while adding records to ' + value + ' collection.')
+              return err;
+            }
+            return;
+          });
+        });
+      } else {
+        await logger.info('Collection ' + value + ' update is SKIPPED');
+      }
+    } else {
+      await logger.info('Collection ' + value + ' is now update');
+      mongoose.connection.db.collection(value, async (err, collection) => {
+        if (err) {
+          await logger.error('FAIL: Error in updating ' + value + ' collection.');
+          return err;
+        }
+        async.each(dbSeedRecords[value], function (obj, asynCb) {
+          collection.insertOne(obj, function (err) {
+            if (err) {
+              logger.error('FAILED while updating "' + value + '" collection. Error: ' + err);
+              return asynCb(err);
+            }
+            return asynCb();
+          });
+        }, function (err) {
+          if (err) {
+            logger.error('FAIL: Error occured while adding records to ' + value + ' collection.')
+            return err;
+          }
+          return;
+        });
+      });
+    }
+  });
+  await logger.info('METHOD EXIT - application.lib.db.updateCollections');
 }
 
 module.exports = {
